Backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
flask==3.0.3
gunicorn==22.0.0
flask-restful==0.3.10
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-limiter==3.7.0
flask-cors==4.0.1
flask-caching==2.3.0

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
alembic==1.13.2

# ─── Async / Concurrency ──────────────────────────────────────────────────────
eventlet==0.36.1
gevent==24.2.1
redis==5.0.8
celery==5.4.0

# ─── AI Integrations ──────────────────────────────────────────────────────────
openai==1.40.0
httpx<0.28.0
tiktoken==0.7.0
pinecone-client==3.2.2
chromadb==0.4.24
sentence-transformers==2.7.0

# ─── Security ─────────────────────────────────────────────────────────────────
cryptography==43.0.0
pyotp==2.9.0
python-dotenv==1.0.1
pydantic==2.8.2
email-validator==2.2.0
bleach==6.1.0

# ─── Utilities ────────────────────────────────────────────────────────────────
marshmallow==3.21.3
python-slugify==8.0.4
pillow==10.4.0
python-magic==0.4.27
humanize==4.10.0
structlog==24.4.0
sentry-sdk[flask]==2.12.0
PyPDF2==3.0.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-flask==1.3.0
pytest-cov==5.0.0
factory-boy==3.3.1
faker==26.0.0
